requirements.txt

pywin32
pytesseract
pdf2image
Pillow
python-dotenv
playwright
pytest
pytest-mock
openpyxl
